Resend leading and IDR frames to make streaming reliable and resumable
Before, these were only sent at the very start. That meant that if you
dropped the start frames (e.g. because the tab was in the background
when they came in, and you didn't have time to handle them) or you
joined later after the stream had already started, then you couldn't do
anything.

With this, they're now stored long-term, and resent at the start of all
connections. This doesn't solve the background loading frame dropping
(watch this space), but you can now just refresh to get a good stream.